requirements.txt
streamlit==1.34.0
python-dotenv==1.0.0
psutil==6.1.0
pytz==2024.2
